Add config option to disable deleting vendor directories
This was an unexpected change!

https://github.com/coenjacobs/mozart/compare/0.4.0...0.5.1#diff-435947b3689be497224d8a16105a5eba

+++ src/Mover.php
@@ -90,7 +90,9 @@ class Mover
             $this->movedPackages[] = $package->config->name;
         }
 
-        $this->deletePackageVendorDirectories();
+        if (!isset($this->config->delete_vendor_directories) || $this->config->delete_vendor_directories === true) {
+	        $this->deletePackageVendorDirectories();
+        }
     }
 
     /**
